refactor: split format/buffers.go model from render (#4661)
Extract the show-system-buffers row/taxonomy MODEL out of buffers.go
into a new buffers_model.go, mirroring the #4656 status_sections.go and
#4657 cos_sections.go model-then-render split. Pure code motion; the
rendered output is byte-identical.

The prior CLI/gRPC/REST buffer-status parity bugs stemmed from the two
renderers building the row set independently: FormatSystemBuffers (text,
used by CLI + gRPC) and StructuredSystemBufferRows (structured, used by
the REST endpoint) each ran the same three-call sample/row/counter
build. This factors that duplicated build into one shared
buildSystemBufferModel returning a systemBufferModel (rows + counterRows
+ knownUMEM/knownTX), so a row or counter added to the taxonomy surfaces
in every renderer at once and cannot drift.

buffers_model.go now owns the taxonomy label constants, the internal
sample/row/counter/binding-key types, buildSystemBufferModel, the row
builders (systemBufferRows / systemBufferCoSRows /
systemBufferFlowCacheAggregate / systemBufferCounterRows), the
WARNING/CRITICAL usage classifier (systemBufferUsage), the helper-status
sampling and per-binding fallback (systemBufferSamples /
applyBindingStatusFallback), and the scope helpers. buffers.go keeps the
two section-heading constants, the exported REST row types, the exported
API (SystemBufferUtilizationRows / StructuredSystemBufferRows + the
exportedSystemBuffer* mappers), and FormatSystemBuffers, now reduced to a
build-then-render orchestrator over two new per-section render helpers
(writeSystemBufferUtilizationSection / writeSystemBufferCountersSection).
The counter-section leading-blank-line spacing (HasSuffix "\n\n") is
preserved. No symbol collision with status_sections.go / cos_sections.go.

Byte-identity gate: TestFormatSystemBuffersGolden pins the full output of
both render modes (detail=false/true) plus the "unavailable" branch for a
fixture that exercises UMEM/TX/CoS aggregate+detail rows, bounded
session/flow-cache/neighbor rows, the WARNING/CRITICAL classes, the
high-utilization warning count, and every status counter with its
per-binding detail rows. The golden fixture was generated by running the
same test on a pristine pre-refactor origin/master worktree; the
refactored code then reproduced it byte-for-byte (test passes without
-update). It reuses the package-level updateGolden flag from
status_golden_test.go.

Validation: go build ./... clean, go vet clean, gofmt clean, go test
./pkg/dataplane/userspace/format/ green. No module-doc change needed: the
format package has no README (matches #4656/#4657, which touched only
_Log.md plus the golden test/fixture), and the split is
behavior-identical.

Closes #4661
