+++ _Log.md
@@ -42254,3 +42254,50 @@ top.
   pkg/daemon/daemon_ha_userspace_stream.go (new),
   pkg/daemon/daemon_ha_userspace_export.go (new),
   pkg/daemon/daemon_ha_userspace_readiness.go (new), _Log.md
+
+## 2026-07-08 — #4661 refactor: split format/buffers.go model from render (773 LOC)
+
+- **Timestamp**: 2026-07-08
+  **Action**: #4661 refactor — split the `show system buffers` row/taxonomy
+  MODEL out of buffers.go into a new buffers_model.go, mirroring the #4656
+  status_sections.go / #4657 cos_sections.go model-then-render pattern. Pure
+  code motion, output byte-identical. The prior CLI/gRPC/REST buffer-status
+  parity bugs came from FormatSystemBuffers (text, used by CLI + gRPC) and
+  StructuredSystemBufferRows (structured, used by the REST endpoint) each
+  independently building the sample/row/counter set; this factored that
+  duplicated three-call build into one shared `buildSystemBufferModel`
+  returning a `systemBufferModel` (rows + counterRows + knownUMEM/knownTX),
+  so a row or counter added to the taxonomy surfaces in every renderer at
+  once and cannot drift. buffers_model.go holds the taxonomy label consts,
+  the internal sample/row/counter/binding-key types, buildSystemBufferModel,
+  systemBufferRows/systemBufferCoSRows/systemBufferFlowCacheAggregate/
+  systemBufferCounterRows (row build), systemBufferUsage (WARNING/CRITICAL
+  classification), systemBufferSamples + applyBindingStatusFallback (helper
+  status sampling + per-binding fallback), and the scope helpers.
+  buffers.go keeps the two section-heading consts, the exported REST row
+  types (SystemBufferUtilizationRow/SystemBufferCounterRow/SystemBufferRows),
+  the exported API (SystemBufferUtilizationRows/StructuredSystemBufferRows +
+  the exportedSystemBuffer* mappers), and FormatSystemBuffers, now reduced to
+  a build-then-render orchestrator over two new per-section render helpers
+  (writeSystemBufferUtilizationSection / writeSystemBufferCountersSection).
+  The counter-section leading-blank-line spacing (`HasSuffix "\n\n"`) is
+  preserved. No symbol collision with status_sections.go / cos_sections.go
+  (checked: distinct writeSystemBuffer* / buildSystemBufferModel names).
+  Byte-identity gate: new TestFormatSystemBuffersGolden pins the full output
+  of both render modes (detail=false/true) plus the "unavailable" branch for
+  a fixture exercising UMEM/TX/CoS aggregate+detail rows, bounded session/
+  flow-cache/neighbor rows, the WARNING/CRITICAL classes, the high-util
+  warning count, and every status counter with per-binding detail. The
+  golden fixture was generated by running the SAME test on a pristine
+  origin/master (pre-refactor) worktree, then the refactored code was
+  verified to reproduce it byte-for-byte (test passes without -update). Reuses
+  the package-level `updateGolden` flag from status_golden_test.go. go build
+  ./... clean, go vet clean, gofmt clean, go test
+  ./pkg/dataplane/userspace/format/ green. No module-doc change needed: the
+  format package has no README (matches #4656/#4657, which touched only
+  _Log.md + the golden test/fixture), and the split is behavior-identical.
+  **File(s)**: pkg/dataplane/userspace/format/buffers.go (trimmed to render +
+  public API), pkg/dataplane/userspace/format/buffers_model.go (new — model +
+  taxonomy + sampling), pkg/dataplane/userspace/format/buffers_golden_test.go
+  (new), pkg/dataplane/userspace/format/testdata/system_buffers.golden (new),
+  _Log.md
